# backend/swagger.yaml
openapi: 3.0.3
info:
  title: SSO Microdata Identity Provider API
  description: |
    ## SSO Microdata — Dokumentasi API & Panduan Integrasi

    Layanan **Single Sign-On (SSO)** berbasis standar **OpenID Connect (OIDC)** yang memungkinkan seluruh aplikasi web Microdata menggunakan satu akun terpusat.

    ### Aplikasi yang Terintegrasi
    | Aplikasi | Framework | Port | Client ID |
    |---|---|---|---|
    | SAIMS (Asset & Inventory) | Next.js + Golang | 3002 | `client-saims` |
    | Microintern (Manajemen PKL) | React + Express | 5173 | `client-microintern` |
    | Arsip Surat | Laravel | 8000 | `client-arsip-surat` |
    | SiteFlow CPC | Laravel | 8001 | `cpc` |

    ### URL Launch Portal (Development Local)
    | Aplikasi | URL Launch |
    |---|---|
    | Arsip Surat | `http://localhost:8000/sso/login` |
    | Microintern PKL | `http://localhost:5173/sso/start` |
    | SiteFlow CPC | `http://localhost:8001/login/sso` |
    | SAIMS Inventaris | `http://localhost:3002/login?sso=1` |

    Browser-facing SSO menggunakan `http://localhost:3000`. Aplikasi di dalam
    container menggunakan `http://host.docker.internal:3000` hanya untuk
    komunikasi server-to-server; issuer dan redirect browser tetap lokal.

    ---

    ### Cara Kerja Alur Login SSO (Authorization Code Flow + PKCE)

    ```
    1. User klik "Masuk dengan SSO"  →  Client App
    2. Client App generate PKCE      →  simpan code_verifier di cookie/session
...
    ```

    ---

...

    #### Prasyarat
...

...
    ```dotenv
...

...
servers:
...
paths:
...
  /oidc/token:
    post:
      tags:
        - OIDC Token & UserInfo
      summary: Exchange Authorization Code untuk Access Token & ID Token
      description: Menukar authorization_code (dengan PKCE) untuk mendapatkan access_token dan id_token.
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required:
                - grant_type
                - client_id
                - code
                - redirect_uri
              properties:
                grant_type:
                  type: string
                  example: authorization_code
                client_id:
                  type: string
                  example: client-saims
                client_secret:
                  type: string
                  example: saims_sso_secret_key_8899aabbccddeeff112233
                code:
                  type: string
                  example: spl3T-8a...
                redirect_uri:
                  type: string
                  example: http://localhost:3002/sso/callback
                code_verifier:
                  type: string
                  example: dBjftJeZ4CVP-mB92K27uhbUJU1p1r_wW1gFWFOEjXk
      responses:
        '200':
          description: Token berhasil terbit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: Invalid authorization code atau PKCE mismatch
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Masukkan Access Token yang didapat dari endpoint /oidc/token
    AdminSessionCookie:
      type: apiKey
      in: cookie
      name: admin_session
      description: Cookie sesi yang diterbitkan oleh POST /admin/login.
  schemas:
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
          example: access-token-placeholder
        token_type:
          type: string
          example: Bearer
        expires_in:
          type: integer
          example: 3600
        id_token:
          type: string
          example: id-token-placeholder
        scope:
          type: string
          example: openid profile email
...